add type conversion when comparing unsigned and signed values

+++ services/audio_policy/server/domain/stream/audio_stream_collector.cpp
@@ -1335,7 +1335,7 @@ void AudioStreamCollector::HandleStartStreamMuteState(StartStreamInfo startStrea
             setStateEvent.streamSetState = StreamSetState::STREAM_MUTE;
             callback->MuteStreamImpl(setStateEvent);
             changeInfo->backMute = true;
-            if (startStreamInfo.sessionId == changeInfo->sessionId) {
+            if (startStreamInfo.sessionId == static_cast<uint32_t>(changeInfo->sessionId)) {
                 silentControl = true;
             }
         } else if (!mute && changeInfo->backMute) {
